Create README.md
My name is Crystal and I am looking to create a record management system to help an adult day company go digital. 
I need help in creating an app for staff to report daily activities and behaviors.
I look forward to learning and finding new ways for things to be done.